+++ api/interfaces/models/IPrimalStakes.cs
@@ -11,6 +11,6 @@ public interface IPrimalStakes
     public string Description { get; }
     public List<string> Keywords { get; }
 
-    public string GetCharacterStageContribution(int seed, string characterStage, IGenre genre, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IDramaticQuestion dramaticQuestion);
+    public string GetCharacterStageContribution(long seed, string characterStage, IGenre genre, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IDramaticQuestion dramaticQuestion);
 
 }
