+++ api/models/dramaticQuestions/OpenCommunication.cs
@@ -14,7 +14,7 @@ public class OpenCommunication : IDramaticQuestion
     public string Negation { get { return "Insantiy"; } }
     public string Positive { get { return "Open Communication"; } }
 
-    public string GetLogLineContribution(int seed, string CompletionType, IGenre genre, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes)
+    public string GetLogLineContribution(long seed, string CompletionType, IGenre genre, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes)
     {
         return $"The overarching theme of this story is asking if babbling nonsense can ultimately communicate an important idea.";
     }
