+++ api/models/sequence/BStory.cs
@@ -10,7 +10,7 @@ public class BStory : ISequence
     public string Name { get { return "B Story"; } }
     public string Description { get { return "The B Story is usually the love story set against the backdrop of the main action."; } }
     
-    public string GetLogLineContribution(int seed, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes, IDramaticQuestion dramaticQuestion)
+    public string GetLogLineContribution(long seed, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes, IDramaticQuestion dramaticQuestion)
     {
         return $"{Name} advice goes here... {Description}";
     }
