+++ api/models/sequence/Debate.cs
@@ -10,7 +10,7 @@ public class Debate : ISequence
     public string Name { get { return "Debate"; } }
     public string Description { get { return "The Hero is in shock and doubts if they have what it takes to tackle the Problem. Having realized stasis=death in the Setup, they are now explicitly confronted with what's at stake for them to lose."; } }
     
-    public string GetLogLineContribution(int seed, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes, IDramaticQuestion dramaticQuestion)
+    public string GetLogLineContribution(long seed, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes, IDramaticQuestion dramaticQuestion)
     {
         return $"{Name} advice goes here... {Description}";
     }
